release: 0.0.8
We've reached feature parity with the old Dorothy. Things are built in a better, more structured way but there is still a lot to abstract and refactor - especially the copy/paste fest that is the GUI.

There's no way to subtract "No Drops" currently and there are some planned UI changes in the next few days.

This was a much more enjoyable experience than using the win32 API.