feat: see description: a_o_t, images, more
Fix always_on_top in local fork of egui and initial image-loading work and god knows what else I lost track.

Persistence needs to be fixed so that I can use RetainedImage for easier ImageButtons and then we are good for next release.

Creating a config fork for the above config changes because there is a huge mess to unfuck.

This commit is mostly to be used as a save state of "Dear god let me go back everything is fucked and it was working before fuck me."