Release: 0.0.13
Some refactoring was in the middle of being done but I want people to be able to read the code of the current release - for security reasons.

Going to see if I can get Microsoft MSVC building in the Github CI so people can trust that the .exe is the same code in the repo.